refactor(repositories): added "repositories" folder to abstract dynamodb transactions from controllers
WIP refactor createUser
