+++ src/controllers/users.js
@@ -1,26 +1,13 @@
-const AWS = require( "aws-sdk" );
-const { v4 } = require( "uuid" );
-const bcrypt = require( 'bcrypt' );
-
-const USERS_TABLE = process.env.USERS_TABLE;
-const dynamoDbClient = new AWS.DynamoDB.DocumentClient();
+const usersRepository = require( '../repositories/usersRepository' );
 
 const getUserByEmail = async ( req, res ) => {
     const { email } = req.params;
-    const params = {
-        TableName: USERS_TABLE,
-        IndexName: 'emailIndex',
-        KeyConditionExpression: 'email = :email',
-        ExpressionAttributeValues: {
-            ':email': email
-        }
-    };
 
     try {
-        const { Items } = await dynamoDbClient.query( params ).promise();
-        if ( Items.length ) {
-            const { userId, name, email } = Items[0];
-            res.json( { success: true, user: { userId, name, email } } );
+        const user = await usersRepository.getUserByEmail( email );
+
+        if ( user ) {
+            res.json( { success: true, user } );
         } else {
             res
                 .status( 404 )
